+++ langchain4j-azure-ai-search/src/main/java/dev/langchain4j/store/embedding/azure/search/AzureAiSearchEmbeddingStore.java
@@ -1,4 +1,4 @@
-package dev.langchain4j.store.embedding.pgvector;
+package dev.langchain4j.store.embedding.azure.search;
 
 import dev.langchain4j.data.embedding.Embedding;
 import dev.langchain4j.data.segment.TextSegment;
